part 4 blog list expansions steps 4.15, 4.16, 4.17, 4.18, 4.19, 4.21 - skipped steps 4.20, 4.22 and 4.23

+++ part-4/blog/controllers/blogPostsController.js
@@ -1,5 +1,15 @@
 const blogRouter = require('express').Router()
 const BlogPost = require('../models/blogPost')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 // Get all blog posts
 
@@ -10,7 +20,7 @@ const BlogPost = require('../models/blogPost')
 // })
 
 blogRouter.get('/', async (request, response) => {
-  const posts = await BlogPost.find({})
+  const posts = await BlogPost.find({}).populate('user', { username: 1, name: 1 })
   response.json(posts)
 })
 
@@ -51,12 +61,21 @@ blogRouter.get('/:id', async (request, response) => { //tiro o param next
 // Create blog post
 blogRouter.post('/', async (request, response) => { //tiro o param next
   const body = request.body
+  // const user = await User.findById(body.user)
+
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
 
   const blogPost = new BlogPost({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
 
   // blogPost.save()
@@ -77,6 +96,8 @@ blogRouter.post('/', async (request, response) => { //tiro o param next
 
   //tirando try catch dps de instalar a lib express-async-errors
   const savedPost = await blogPost.save()
+  user.blogPosts = user.blogPosts.concat(savedPost._id)
+  await user.save()
   response.status(201).json(savedPost)
 })
 
@@ -97,8 +118,17 @@ blogRouter.delete('/:id', async (request, response) => { //tiro o param next
   //   next(exception)
   // }
 
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await BlogPost.findById(request.params.id)
+  const userFromBlog = blog.user.toString()
+
+  if(userFromBlog === user._id.toString())
   //tirando try catch dps de instalar a lib express-async-errors
-  await BlogPost.findByIdAndDelete(request.params.id)
+    await BlogPost.findByIdAndDelete(request.params.id)
   response.status(204).end()
   //Because of the library, we do not need the next(exception) call anymore. The library handles everything under the hood. If an exception occurs in an async route, the execution is automatically passed to the error handling middleware.
 })
